extend bar chart
