+++ src/components/charts/lines/LineChartDefaultPro/index.tsx
@@ -12,7 +12,7 @@ import { ChartGranularitySelectField } from '../../shared/ChartGranularitySelect
 
 export type LineChartProPropsOnLineClicked = { axisDimensionValue: string | null };
 
-export type LineChartProProp = {
+export type LineChartProProps = {
   xAxis: Dimension;
   measures: Measure[];
   results: DataResponse;
@@ -29,7 +29,7 @@ export type LineChartProProp = {
   onLineClicked?: LineChartProOptionsClick;
 } & ChartCardHeaderProps;
 
-const LineChartPro = (props: LineChartProProp) => {
+const LineChartPro = (props: LineChartProProps) => {
   const theme: Theme = useTheme() as Theme;
   i18nSetup(theme);
