+++ src/components/charts/utils/granularity.utils.ts
@@ -1,6 +1,9 @@
 import { Dimension, Granularity } from '@embeddable.com/core';
 
-export const getDimensionWithGranularity = (dimension: Dimension, granularity?: Granularity) => {
+export const getDimensionWithGranularity = (
+  dimension: Dimension,
+  granularity?: Granularity,
+): Dimension => {
   const currentGranularity = granularity ?? dimension.inputs?.granularity;
 
   return {
